"Fixed" the world size.
The camera was set to to large a viewing angle (48x32 meters) which was
causing everything to fall "slowly".  I made a few adjustments to that.

Note that there are a ton of magic numbers in this and I'm going to need
to figure out where they're coming from to fix them.